nib file updates.
I noticed that the NSTextView containing the graphing code
was not responding to the return key. To address this, I
explored the nib file and found that unchecking the "Field
Editor" box resolved the problem.  Since we are using it
to edit code, I also unchecked the "Rich Text" and "Font
Panel" boxes. The same changes were made for both the mixed
and the all-Nu example.